+++ src/test/java/pl/szczesniak/dominik/whattowatch/movies/RemoveMovieToWatchFromListRestInvoker.java
@@ -16,7 +16,7 @@ public class RemoveMovieToWatchFromListRestInvoker {
 
 	private final TestRestTemplate restTemplate;
 
-	public ResponseEntity<Void> removeMovie(final Integer userId, final Integer movieId) {
+	public ResponseEntity<Void> removeMovie(final Long userId, final Integer movieId) {
 		final HttpHeaders headers = new HttpHeaders();
 		headers.set("userId", String.valueOf(userId));
 		return restTemplate.exchange(
